requirements_test.txt
# Home Assistant testing requirements
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0

# Component dependencies
aiohttp>=3.8.0
voluptuous>=0.13.0

# Additional testing utilities
pytest-cov>=4.0.0
pytest-timeout>=2.1.0